refactor into library
