+++ Makefile
@@ -2,6 +2,8 @@ UNIXOS = $(shell uname -s)
 
 all : cursbench
 
+cursbench : cursbench.o curslib.o
+
 ifeq ($(UNIXOS),Linux)
   include Makefile.linux
 endif
@@ -9,3 +11,9 @@ endif
 ifeq ($(UNIXOS),Darwin)
   include Makefile.osx
 endif
+
+clean :
+	-rm cursbench
+	-rm *.o
+	-rm *~
+
