+++ Makefile.linux
@@ -3,6 +3,3 @@
 # -lm : required by ceil(3)
 CFLAGS = -Wall -O2 -std=c99 -D_XOPEN_SOURCE -D_DEFAULT_SOURCE
 LDLIBS = -lm -lncurses
-
-clean :
-	-rm cursbench
