+++ Makefile.osx
@@ -1,5 +1,2 @@
-CFLAGS = -Wall -O2 -std=c99 -lncurses
-
-clean :
-	-rm *~ cursbench
-	-rm -df *.dSYM
+CFLAGS = -Wall -O2 -std=c99
+LDLIBS = -lncurses
