+++ curslib.h
@@ -1,0 +1,11 @@
+void drawbar(double frac, int width, int line, int offset);
+
+void drawline(int row, int width);
+
+void write_matrix(int nrows, int ncols, int docolor);
+
+void display_mbps(int dk, int nrows, int ncols, int docolor);
+
+void static_display(char* str, int nrows, int ncols);
+
+void init_colors();
